+++ R/build-home-index.R
@@ -54,9 +54,11 @@ data_home <- function(pkg = ".") {
   pkg <- as_pkgdown(pkg)
 
   print_yaml(list(
-    pagetitle = pkg$meta$home[["title"]] %||% pkg$desc$get("Title")[[1]],
+    pagetitle = pkg$meta$home[["title"]] %||%
+      uncranquote(pkg$desc$get("Title")[[1]]),
     sidebar = data_home_sidebar(pkg),
-    opengraph = list(description = pkg$meta$home[["description"]] %||% pkg$desc$get("Description")[[1]])
+    opengraph = list(description = pkg$meta$home[["description"]] %||%
+                       uncranquote(pkg$desc$get("Description")[[1]]))
   ))
 }